Add tests for 4C beam potential interactions
