+++ tests/reference-files/test_four_c_simulation_beam_potential_helix_reference.dat
@@ -1,0 +1,64 @@
+// -----------------------------------------------------------------------------
+// This input file was created with MeshPy.
+// Copyright (c) 2018-2024
+//     Ivo Steinbrecher
+//     Institute for Mathematics and Computer-Based Simulation
+//     Universitaet der Bundeswehr Muenchen
+//     https://www.unibw.de/imcs-en
+// -----------------------------------------------------------------------------
+------------------------------------------------------------------BEAM POTENTIAL
+POT_LAW_PREFACTOR                     -0.001 1.245e-07
+POT_LAW_EXPONENT                      6.0 12.0
+BEAMPOTENTIAL_TYPE                    Volume
+CUTOFF_RADIUS                         10.0
+STRATEGY                              SingleLengthSpecific_SmallSepApprox_Simple
+NUM_INTEGRATION_SEGMENTS              1
+NUM_GAUSSPOINTS                       10
+POTENTIAL_REDUCTION_LENGTH            15.0
+AUTOMATIC_DIFFERENTIATION             False
+-----------------------------------------------BEAM POTENTIAL/RUNTIME VTK OUTPUT
+VTK_OUTPUT_BEAM_POTENTIAL             yes
+INTERVAL_STEPS                        yes
+EVERY_ITERATION                       yes
+FORCES                                yes
+MOMENTS                               yes
+WRITE_FORCE_MOMENT_PER_ELEMENTPAIR    yes
+-----------------------------------------------------------------------MATERIALS
+MAT 1 MAT_BeamReissnerElastHyper YOUNG 1000 POISSONRATIO 0.0 DENS 0.0 CROSSAREA 0.7853981633974483 SHEARCORR 1.0 MOMINPOL 0.09817477042468103 MOMIN2 0.04908738521234052 MOMIN3 0.04908738521234052
+--------------------------------------------------------------------------FUNCT1
+COMPONENT 0 SYMBOLIC_FUNCTION_OF_SPACE_TIME t
+--------------------------------------------------DESIGN POINT DIRICH CONDITIONS
+DPOINT 1
+E 1 - NUMDOF 9 ONOFF 1 1 1 1 1 1 0 0 0
+------------------------------------DESIGN LINE BEAM POTENTIAL CHARGE CONDITIONS
+DLINE 2
+E 1 - POTLAW 1 VAL 1.0 FUNCT 1
+E 1 - POTLAW 2 VAL 2.0 FUNCT none
+-------------------------------------------------------------DNODE-NODE TOPOLOGY
+NODE 1 DNODE 1
+-------------------------------------------------------------DLINE-NODE TOPOLOGY
+NODE 1 DLINE 1
+NODE 2 DLINE 1
+NODE 3 DLINE 1
+NODE 4 DLINE 1
+NODE 5 DLINE 1
+NODE 6 DLINE 1
+NODE 7 DLINE 1
+NODE 8 DLINE 1
+NODE 9 DLINE 1
+---------------------------------------------------------------------NODE COORDS
+NODE 1 COORD 2 0 0
+NODE 2 COORD 1.62192623901 1.17019454588 1.25
+NODE 3 COORD 0.630644724791 1.89796923871 2.5
+NODE 4 COORD -0.599067012379 1.90817156322 3.75
+NODE 5 COORD -1.60228723109 1.19694428821 5
+NODE 6 COORD -1.99972469016 0.0331837844587 6.25
+NODE 7 COORD -1.64111871468 -1.14312263748 7.5
+NODE 8 COORD -0.662048814506 -1.8872443846 8.75
+NODE 9 COORD 0.567324370926 -1.91784854933 10
+--------------------------------------------------------------STRUCTURE ELEMENTS
+1 BEAM3R HERM2LINE3 1 3 2 MAT 1 TRIADS -0.613943125569 -1.48218982027 0.613943125569 0.417368945844 -1.76423841509 1.5426293588 -0.134747911956 -1.67656400113 1.090228612
+2 BEAM3R HERM2LINE3 3 5 4 MAT 1 TRIADS 0.417368945844 -1.76423841509 1.5426293588 -1.62734522345 1.40867035905 -2.14678720617 1.04207424493 -1.71966690312 1.95284589367
+3 BEAM3R HERM2LINE3 5 7 6 MAT 1 TRIADS -1.62734522345 1.40867035905 -2.14678720617 -1.73850502121 0.154252595589 -1.33838174598 -1.7587433354 0.745649232387 -1.77087296744
+4 BEAM3R HERM2LINE3 7 9 8 MAT 1 TRIADS -1.73850502121 0.154252595589 -1.33838174598 -1.36158461459 -0.809718268376 -0.390228435716 -1.59922161847 -0.364347553341 -0.873187825362
+-----------------------------------------------------------------------------END
